benchmark/pyproject.toml
[project]
name = "benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "psycopg[binary]>=3.3.1",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "locust>=2.42.6",
]
